+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # CasualOS Changelog
 
+## V1.1.6
+
+### Date: 6/11/2020
+
+### Changes:
+
+-   :boom: Breaking Changes
+
+    -   Renamed all the history tags to not have the `aux` prefix.
+
+-   :rocket: Improvements
+
+    -   Added the `server.storyPlayerCount()` function.
+        -   Returns a promise that resolves with the number of players currently connected to the current story.
+        -   Optionally accepts a parameter which indicates the story to check.
+    -   Added the `server.totalPlayerCount()` function.
+        -   Returns a promise that resolves with the total number of players connected to the server.
+    -   Added the `server.stories()` function.
+        -   Returns a promise that resolves with the list of stories that are on the server.
+
+-   :bug: Bug Fixes
+    -   Removed the globals bot tags from the documentation since they no longer exist.
+
 ## V1.1.5
 
 ### Date: 6/9/2020